+++ packages/@eventual/compiler/src/eventual-infer.ts
@@ -8,7 +8,7 @@ import { generateSchema } from "@anatine/zod-openapi";
 import {
   commands,
   CommandSpec,
-  eventHandlers,
+  subscriptions,
   SubscriptionSpec,
   events,
   ServiceSpec,
@@ -31,7 +31,12 @@ import esbuild from "esbuild";
 import fs from "fs/promises";
 import os from "os";
 import path from "path";
-import { getSpan, isCommandCall, isOnEventCall } from "./ast-util.js";
+import {
+  getSpan,
+  isCommandCall,
+  isOnEventCall,
+  isSubscriptionCall,
+} from "./ast-util.js";
 import { printModule } from "./print-module.js";
 
 export async function infer(
@@ -85,13 +90,17 @@ export async function infer(
       )
     ),
     subscriptions: Object.fromEntries(
-      eventHandlers().map((e) => [
+      subscriptions().map((e) => [
         e.name,
         {
           name: e.name,
-          runtimeProps: e.runtimeProps,
+          props: {
+            memorySize: e.props?.memorySize,
+            retryAttempts: e.props?.retryAttempts,
+            timeout: e.props?.timeout,
+          },
           sourceLocation: e.sourceLocation,
-          subscriptions: e.subscriptions,
+          filters: e.filters,
         } satisfies SubscriptionSpec,
       ])
     ) as ServiceSpec["subscriptions"],
@@ -107,6 +116,10 @@ export async function infer(
           method: command.method,
           input: command.input ? generateSchema(command.input) : undefined,
           output: command.output ? generateSchema(command.output) : undefined,
+          passThrough: command.passThrough,
+          params: command.params,
+          validate: command.validate,
+          internal: command.internal,
         } satisfies CommandSpec,
       ])
     ),
@@ -190,7 +203,10 @@ export class InferVisitor extends Visitor {
   }
 
   visitCallExpression(call: CallExpression): Expression {
-    if (this.exportName && (isCommandCall(call) || isOnEventCall(call))) {
+    if (
+      this.exportName &&
+      (isCommandCall(call) || isOnEventCall(call) || isSubscriptionCall(call))
+    ) {
       this.didMutate = true;
 
       return {